V3 make use of Reserved32 with SectorSize.

+++ inc/winfsp/winfsp.hpp
@@ -492,11 +492,11 @@ public:
     }
 
     /* properties */
-    INT16 SectorSize()
+    UINT32 SectorSize()
     {
         return _VolumeParams.SectorSize;
     }
-    VOID SetSectorSize(INT16 SectorSize)
+    VOID SetSectorSize(UINT32 SectorSize)
     {
         _VolumeParams.SectorSize = SectorSize;
     }
